Safety Dance!

+++ get_template.sh
@@ -4,34 +4,39 @@
 read -p "What do you want to call it? > " projectName
 
 git clone https://github.com/dotHTM/template_jekyll.git "$projectName"
-cd template_jekyll
-rm -rf .git
-rm get_template.sh
-git init
-git add *
 
-# --- # --- # --- # --- # --- # --- # --- # --- 
-# Open in Editor
 
-editor=''
-if [[ -e `which subl` ]]; then
-    editor='subl'
-fi
+if [[ -e "$projectName/.git" ]]; then
+    #statements
+    cd "$projectName"
+    rm -rf .git
+    rm get_template.sh
+    git init
+    git add *
 
-if [[ -n $editor ]]; then
-    $editor ./
-    $editor ./README.md
-fi
+    # --- # --- # --- # --- # --- # --- # --- # --- 
+    # Open in Editor
 
-# --- # --- # --- # --- # --- # --- # --- # --- 
-# Open in Git helper
+    editor=''
+    if [[ -e `which subl` ]]; then
+        editor='subl'
+    fi
 
-gitHelper=''
-if [[ -e `which smerge` ]]; then
-    gitHelper='smerge'
-fi
+    if [[ -n $editor ]]; then
+        $editor ./
+        $editor ./README.md
+    fi
 
-if [[ -n $gitHelper ]]; then
-    $gitHelper ./
+    # --- # --- # --- # --- # --- # --- # --- # --- 
+    # Open in Git helper
+
+    gitHelper=''
+    if [[ -e `which smerge` ]]; then
+        gitHelper='smerge'
+    fi
+
+    if [[ -n $gitHelper ]]; then
+        $gitHelper ./
+    fi
 fi
